// Frontend/template/demo_1/src/app/models/paciente.model.ts
export interface Paciente {
  _id: string
  nombre: string
  apellidoPaterno: string
  apellidoMaterno: string
  dni: string
  celular: string
  email: string
  fechaNaciemineto: Date
  direccion: string
  estado: number
  nombreFamiliar: string
  dniFamiliar: string
  parentesco: string
  celularFamiliar: string
}